backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.22.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0